# src/app/api/openapi.yaml
swagger: "2.0"
info:
  title: "WorkUP API"
  description: "API Documentation"
  version: "2.0.0"
host: "prod-team-33-pafku5n2.final.prodcontest.ru"
basePath: "/api"
schemes:
  - "https"
  - "http"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Введите 'Bearer ' перед вашим токеном, например: 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'"
paths:
  /auth/reg/:
    post:
      tags:
        - auth
      summary: "Регистрация пользователя"
      description: "Эндпоинт для регистрации нового пользователя."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Объект пользователя для регистрации. Обязательные поля: name, surname, password, login; опционально: tags, description, job, company."
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
                example: "Иван"
              surname:
                type: "string"
                example: "Иванов"
              password:
                type: "string"
                example: "Passw0rd!"
              login:
                type: "string"
                example: "ivanivanov"
              tags:
                type: "array"
                items:
                  type: "string"
                example: ["developer", "mentor"]
              description:
                type: "string"
                example: "Новый пользователь платформы WorkUP"
              job:
                type: "string"
                example: "Разработчик"
              company:
                type: "string"
                example: "ООО Рога и Копыта"
            required:
              - name
              - surname
              - password
              - login
      responses:
        200:
          description: "Пользователь успешно зарегистрирован. Возвращается JWT-токен."
          schema:
            $ref: "#/definitions/TokenResponse"
        400:
          description: "Ошибка валидации входных данных"
          schema:
            type: "object"
            properties:
              error:
                type: "array"
                items:
                  type: "string"
                example: ["Поле login обязательно"]
        409:
          description: "Пользователь с таким логином уже существует"
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Пользователь с данным логином уже зарегистрирован"
  /auth/login/:
    post:
      tags:
        - auth
      summary: "Аутентификация пользователя"
      description: "Эндпоинт для получения JWT-токена по логину и паролю."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Учетные данные пользователя для аутентификации"
          required: true
          schema:
            type: "object"
            properties:
              login:
                type: "string"
                example: "ivanivanov"
              password:
                type: "string"
                example: "Passw0rd!"
            required:
              - login
              - password
      responses:
        200:
          description: "JWT-токен успешно выдан"
          schema:
            type: "object"
            properties:
              token:
                type: "string"
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        401:
          description: "Неверные учетные данные"
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Неверный логин или пароль"
  /auth/:
    patch:
      tags:
        - auth
      summary: "Обновление информации о пользователе"
      description: "Эндпоинт для обновления информации о пользователе. Все поля являются необязательными. Пользователь обновляет данные на основе JWT-токена."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          description: "Объект с обновляемыми данными пользователя. Допустимые поля: name, surname, password, description, job, company."
          required: false
          schema:
            type: "object"
            properties:
              name:
                type: "string"
                example: "Иван"
              surname:
                type: "string"
                example: "Иванов"
              password:
                type: "string"
                example: "Passw0rd!"
              description:
                type: "string"
                example: "Обновлённое описание пользователя"
              job:
                type: "string"
                example: "Разработчик"
              company:
                type: "string"
                example: "ООО Рога и Копыта"
      responses:
        200:
          description: "Пользователь успешно обновлён"
          schema:
            $ref: "#/definitions/StatusResponse"
        401:
          description: "Пользователь не найден или ошибка авторизации"
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Ошибка авторизации: пользователь не найден"
...
definitions:
  TokenResponse:
    type: object
    properties:
      status:
        type: string
        example: "ok"
      token:
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
  StatusResponse:
    type: object
    properties:
      status:
        type: string
        example: "ok"
...